requirements.txt
qrcode
configparser
pytelegrambotapi
pillow
